Render filled form fields into Glyphs
Now form data will be rendered as usual along with the rest of the
content into the output.

This changes the page's Glyphs structure to include filled form field
characters and their bounding boxes.

I have tested this locally on a few PDFs. We should test it a little
more.