Switch to sensiblecodeio fork of msgpack-cpp
This is so we can incorporate changes needed for a new pdf2msgpack
release.

The branch used has changes made in msgpack-cpp 4.1.2 reverted.
This is because those changes result in floats and doubles getting
packed as ints when the value after the decimal is zero.
This breaks our current promise that values are floats.

Alternative fixes might be:

* change the specification, and downstream consumers have to convert
  ints to floats
* overload the packer to handle floats and doubles in the previous way

The current changes to msgpack-cpp are clean and do not require any other
changes beyond reverting the relevant commits. The net effect is that
the packing code for doubles and floats is exactly as it was in 4.1.1.