Update Poppler to use sensiblecodeio fork
This is so we can apply a revert on this change:

https://gitlab.freedesktop.org/poppler/poppler/-/merge_requests/823

for the time being.

It affects outputs of right-to-left text, and for some reason removes
the spacing between words. It should be investigated in future, but for
now, working around the problem, so that we can continue updating
Poppler.

The filepath has been left unchanged, only the submodule URL has been
updated. This is for convenience, to avoid having to change the
Dockerfile, and so we can more simply revert back to upstream in future.